v2.2.0 Release Prep
This includes new semantics around the RetNullable type, around dealing with contained null values in type safe ways.

The 'result()' call adds a new overload method that returns a default value if the stored value is null.  The corresponding 'getValue()' is not overloaded, because there are specific semantics around the null return value, and problem state would still need a null value returned.

The 'requireNonNull()' method adds support to return a RetVal only if the stored value is non-null, and return problems if null.

The 'defaultAs()' call returns a RetVal with a default value if the stored value is null, and the value as-is if non-null or a problem.

The 'consumeIfNonnull()' returns a RetVoid, and consumes the value if non-null and not a problem.

The 'defaultOrMap()' acts like a 'map()' call, but returns a default value on null values.

The 'nullOrMap()' returns a RetNullable, with a specific 'map()' if the stored value is non-null.

The 'nullOrThenNullable()' is the then equivalent of nullOrMap().

The gradle version was updated to 7.4.2.